Signal System Updates
- Modified receive/sending system to accommodate larger requests
