--- airport.cs
@@ -1,15 +1,0 @@
-using System;
-using plane;
-using controls;
-
-namespace airport
-{
-  public class Hangar
-  {
-    public bool Full = false;
-    public Hangar(string HangarCode, int HangarSize)
-    {
-      
-    }
-  }
-}
